feat: 역할 분리 및 이동
역할 분리와 이동
- ApplicationMain : 클라이언트
- RacingGame : 컨트롤러
- Cars : 모델
