+++ README.md
@@ -29,15 +29,16 @@
   - [x] 자동차의 위치를 출력할 수 있다.
   - [ ] 자동차 이름을 출력한다.
 - 레이싱 게임
-  - [x] 자동차들을 움직인다.
-  - [ ] 우승자를 알려준다.
-- 게임 컨트롤러
   - [x] 입력을 받고 게임을 진행한뒤 결과를 출력한다.
 - 닉네임
   - [x] 닉네임은 null일 수 없다.
   - [x] 닉네임 앞뒤에 white space가 있으면 white space는 제거된다.
   - [x] 닉네임은 공백일 수 없다.
   - [x] 닉네임은 5자를 초과할 수 없다.
+- 레이싱 출전 자동차
+  - [ ] 출전될 자동차를 세팅한다.
+  - [ ] 자동차들을 움직인다.
+  - [ ] 자동차들의 위치상태를 알려준다.
 
 ### 프로그래밍 요구사항
 - 모든 로직에 단위 테스트를 구현한다. 단, UI(System.out, System.in) 로직은 제외.
