+++ src/main/java/racingcar/ApplicationMain.java
@@ -1,0 +1,9 @@
+package racingcar;
+
+public class ApplicationMain {
+
+    public static void main(String[] args) {
+        RacingGame racingGame = new RacingGame();
+        racingGame.start();
+    }
+}
